added menu background image, removed unnecessary function
